Committed by patch #2219361 ("BS1: Simplified subtitles"). It uses the lockScreen() / unlockScreen() backend API, instead of copyRectToScreen(). Before, it had to copy a piece of the engine's screen to draw on to simulate transparency, which was awkward.
However, this means we're now forcing full-screen updates on each frame. There
may be performance regressions, particularly if the frames are much smaller
than the screen. Hopefully, it's the decoding that's the bottleneck, but if
this causes problems on low-end devices... well, Fingolfin had some ideas about
that when he first proposed the lockScreen() / unlockScreen() API.

svn-id: r35115