SCI: Remove remarks about in-place loading, at least for now.
Not loading in-place has the advantage that it allows to cleanly
cancel loading a savesate, e.g. if we detect a defect during
loading. This is much harder (if not impossible) to achieve if
loading in-place. So I'd prefer if we took advantage of this.
Right now, though, we don't.

svn-id: r45786