Fix a valgrind warning.
It is *not* a good idea to pass a reference to a list entry to List::remove.
Since List::remove will remove *all* occurances of that list entry, it will
also invaldiate the reference, resulting in invalid memory reads after the
entry has been removed from the list, when List::remove will continue to
search the rest of the list for more occurances of the same entry.

svn-id: r48225