Add dists/redhat/scummvm-tools.spec.in, and add it to tools/update-version.pl
This is not quite perfect because in principle the ScummVM Tools version
might differ from the ScummVM version. But this way we are at least
reminded during a release that we may have to adjust this.

A potentially better approach would be to move scummvm-tools.spec to the tools
SVN module, and adding a similar update-version.pl script there.

svn-id: r48454