AGI: Fix bug #2862508.
Bug #2862508: "AGI: Black cauldron save through GMM". BC does not
have input line, and that was used for determining that user
can save. Added special case for BC, so it is always allowed to
save. In fact, original does not allow saving in some rare
occasions, but that will require analysing of variables, which
is a gross hack.

svn-id: r49738