BUILD: Unify how MODULE_DIRS is computed for backends
This should help avoid situations where MODULE_DIRS is not set to a
complete list of build dirs (which causes troubles with the automatic
header dependency detection logic).

On the long run, we should replace the relevant code by a macro or also
use rules.mk for this (with yet another if/else case add to it).

svn-id: r51467