SCI: Fix uninitialized variable warning in g++.
Thanks to clone2727 for suggestion to initialize the default new plane
priority to 0xFFFF, which means that the plane should be invisible. I added a
FIXME there though. So please re check this change. In case that is fine just
remove the FIXME, in case some other default value should be chosen, please
change the value and also remove the FIXME.

svn-id: r51561