PSP: backported from trunk
PSP: moved VramAllocator to display_manager.cpp.

It didn't really belong in memory.cpp and we're going to want to include memory.h everywhere.
* * *
PSP: more Vram Allocator cleanup
* * *
PSP: Split up PspMemory class.

PspMemorySwap is more specific ie. it only needs to be known by PSP files. It could be put in another file, but not worth the effort.
* * *
PSP: switch from wrapping memcpy to defining our own memcpy

The advantage is that we get to do inlining and even use lwl and lwr instructions where appropriate. We have to do it ourselves because the PSP doesn't tolerate built-in instructions, but also we have a more efficient memcpy than the lib's.

svn-id: r52873