IPHONE: Enable cruise & teenagent; disable sci32
Actually, currently all engines are added to the project, but not all
are activated. Enabling and disabling engines is currently not easily
possible from within the project itself. So for now I manually enabled
exactly those engines which configure enables by default.

A proper solution would be to refactor the project, with (sub)targets
resp. (sub)projects for every engine. We might also want to switch the
iPhone port to use dynamic plugins.

svn-id: r53138