DC: Ensure creation of output sections .{c,d}tors
Add an object with an empty .ctors and .dtors section to the build
of each plugin.  This ensures that there will be an output section
.ctors and one .dtors in the output file (they could be empty though).
If these output sections were to be removed, the symbols defined inside
by the linker script could disappear as well, and we don't want that.

svn-id: r55149